25 Maret 2025

+++ database/migrations/2025_03_12_042720_create_barangs_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             
             $table->string('name');
             $table->text('description')->nullable();
-            $table->integer('jumlah');
+            $table->decimal('jumlah', 12, 3);
             $table->string('satuan')->nullable();
             $table->string('image')->nullable();
